feat: add Create a New DAO button
