+++ frontend/src/pages/Landing/index.tsx
@@ -1,7 +1,7 @@
 import { Grid, SxProps, Theme } from '@mui/material'
 import { useEffect, useState } from 'react'
 import ContentBlock from '../../components/ContentBlock'
-import MyDAOs from '../../components/MyDAOs'
+import MyDAOs from '../Dao/components/MyDAOs'
 import { PageLayout } from '../../layout/Page'
 
 export default function Landing() {
@@ -26,24 +26,9 @@ export default function Landing() {
       <ContentBlock>
         <h1>The Launchpad of For-Profit DAOs</h1>
       </ContentBlock>
-
-      <Grid container>
-        <Grid item sm={6}>
-          <ContentBlock sx={{ ...blockStyle }} title="Your DAOs">
-            <MyDAOs />
-          </ContentBlock>
-        </Grid>
-        <Grid item sm={6}>
-          <ContentBlock sx={{ ...blockStyle }} title="Create a for-profit DAO">
-            <p>
-              <span>
-                <a href="https://app.kali.gg/">Create a new for-profit Kali DAO</a> with legal benefits. Make sure to
-                choose Company Series LLC template! Then return here to manage sweat equity projects and more.
-              </span>
-            </p>
-          </ContentBlock>
-        </Grid>
-      </Grid>
+      <ContentBlock title="Your DAOs" cta={{ href: 'create-dao', text: 'Create a new DAO' }}>
+        <MyDAOs />
+      </ContentBlock>
     </PageLayout>
   )
 }
